# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CloudLens API",
    "description": "Azure FinOps Managed Service \u2014 cost intelligence API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/tenants/": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "List Tenants",
        "description": "List all tenant configurations.",
        "operationId": "list_tenants_api_v1_tenants__get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TenantConfig"
                  },
                  "title": "Response List Tenants Api V1 Tenants  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Create Tenant",
        "description": "Create a new tenant configuration.\nSP credentials are stored in Key Vault; only the secret reference is persisted.",
        "operationId": "create_tenant_api_v1_tenants__post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantConfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "Get Tenant",
        "operationId": "get_tenant_api_v1_tenants__tenant_id__get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantConfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "tenants"
        ],
        "summary": "Update Tenant",
        "operationId": "update_tenant_api_v1_tenants__tenant_id__patch",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantConfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tenants"
        ],
        "summary": "Delete Tenant",
        "description": "Soft-delete: sets active=False. Does not remove data.",
        "operationId": "delete_tenant_api_v1_tenants__tenant_id__delete",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/waste/{tenant_id}": {
      "get": {
        "tags": [
          "waste"
        ],
        "summary": "List Waste Items",
        "description": "Return waste items for a tenant, optionally filtered by priority.",
        "operationId": "list_waste_items_api_v1_waste__tenant_id__get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Priority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "name": "resolved",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include resolved items",
              "default": false,
              "title": "Resolved"
            },
            "description": "Include resolved items"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WasteItem"
                  },
                  "title": "Response List Waste Items Api V1 Waste  Tenant Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/waste/{item_id}/resolve": {
      "patch": {
        "tags": [
          "waste"
        ],
        "summary": "Resolve Waste Item",
        "description": "Mark a waste item as resolved.",
        "operationId": "resolve_waste_item_api_v1_waste__item_id__resolve_patch",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WasteResolve"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WasteItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanTier": {
        "type": "string",
        "enum": [
          "starter",
          "growth",
          "enterprise"
        ],
        "title": "PlanTier"
      },
      "Priority": {
        "type": "string",
        "enum": [
          "critical",
          "high",
          "medium",
          "low"
        ],
        "title": "Priority"
      },
...
      "TenantConfig": {
        "properties": {
          "tenant_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Tenant Name",
            "description": "Display name of the customer"
          },
          "subscription_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Subscription Ids",
            "description": "Azure subscription IDs to monitor"
          },
          "plan_tier": {
            "$ref": "#/components/schemas/PlanTier",
            "default": "growth"
          },
          "alert_email": {
            "type": "string",
            "title": "Alert Email",
            "description": "Recipient of weekly digest and alerts"
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": true
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Cosmos discriminator",
            "default": "tenant"
          },
          "sp_secret_ref": {
            "type": "string",
            "title": "Sp Secret Ref",
            "description": "Key Vault secret name for SP credentials"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "last_ingested_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Ingested At"
          },
          "last_ingest_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Ingest Error"
          }
        },
        "type": "object",
        "required": [
          "tenant_name",
          "subscription_ids",
          "alert_email",
          "sp_secret_ref"
        ],
        "title": "TenantConfig",
        "description": "Cosmos DB document \u2014 partition key: id"
      },
      "TenantCreate": {
        "properties": {
          "tenant_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Tenant Name",
            "description": "Display name of the customer"
          },
          "subscription_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Subscription Ids",
            "description": "Azure subscription IDs to monitor"
          },
          "plan_tier": {
            "$ref": "#/components/schemas/PlanTier",
            "default": "growth"
          },
          "alert_email": {
            "type": "string",
            "title": "Alert Email",
            "description": "Recipient of weekly digest and alerts"
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": true
          },
          "sp_client_id": {
            "type": "string",
            "title": "Sp Client Id",
            "description": "Service principal client ID"
          },
          "sp_client_secret": {
            "type": "string",
            "title": "Sp Client Secret",
            "description": "Service principal secret (stored to Key Vault)"
          },
          "sp_tenant_id": {
            "type": "string",
            "title": "Sp Tenant Id",
            "description": "Customer Azure AD tenant ID"
          }
        },
        "type": "object",
        "required": [
          "tenant_name",
          "subscription_ids",
          "alert_email",
          "sp_client_id",
          "sp_client_secret",
          "sp_tenant_id"
        ],
        "title": "TenantCreate",
        "description": "Payload for POST /api/v1/tenants"
      },
      "TenantUpdate": {
        "properties": {
          "tenant_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Name"
          },
          "subscription_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Ids"
          },
          "plan_tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanTier"
              },
              {
                "type": "null"
              }
            ]
          },
          "alert_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Email"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "type": "object",
        "title": "TenantUpdate",
        "description": "Payload for PATCH /api/v1/tenants/{id} \u2014 all fields optional"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WasteItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "waste_item"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "subscription_id": {
            "type": "string",
            "title": "Subscription Id"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "resource_name": {
            "type": "string",
            "title": "Resource Name"
          },
          "resource_group": {
            "type": "string",
            "title": "Resource Group"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type",
            "default": ""
          },
          "waste_type": {
            "$ref": "#/components/schemas/WasteType"
          },
          "monthly_cost_eur": {
            "type": "number",
            "minimum": 0.0,
            "title": "Monthly Cost Eur",
            "description": "Current monthly cost of this resource"
          },
          "saving_eur": {
            "type": "number",
            "minimum": 0.0,
            "title": "Saving Eur",
            "description": "Estimated saving if remediated"
          },
          "saving_pct": {
            "type": "number",
            "title": "Saving Pct",
            "description": "Saving as % of current cost",
            "default": 0.0
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation",
            "description": "Human-readable action (EN)"
          },
          "recommendation_it": {
            "type": "string",
            "title": "Recommendation It",
            "description": "Raccomandazione in italiano"
          },
          "advisor_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisor Ref",
            "description": "Azure Advisor recommendation ID"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence",
            "description": "Supporting metrics: cpu_avg, disk_state, etc."
          },
          "detected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Detected At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "snoozed_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snoozed Until"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "subscription_id",
          "resource_id",
          "resource_name",
          "resource_group",
          "waste_type",
          "monthly_cost_eur",
          "saving_eur",
          "priority",
          "recommendation",
          "recommendation_it"
        ],
        "title": "WasteItem",
        "description": "Detected waste item \u2014 Cosmos container: waste_items"
      },
      "WasteResolve": {
        "properties": {
          "resolved_by": {
            "type": "string",
            "minLength": 1,
            "title": "Resolved By"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "resolved_by"
        ],
        "title": "WasteResolve"
      },
      "WasteType": {
        "type": "string",
        "enum": [
          "idle_vm",
          "unattached_disk",
          "orphan_public_ip",
          "oversized_vm",
          "dev_test_eligible",
          "reserved_instance",
          "idle_app_service",
          "unused_load_balancer",
          "old_snapshots",
          "cold_storage",
          "duplicated_backup",
          "expired_cert"
        ],
        "title": "WasteType"
      }
    }
  }
}